third_party/WebKit/LayoutTests/inspector/elements/styles-4/styles-source-lines-inline-expected.txt

Tests that elements panel shows proper inline style locations in the sidebar panel.


[expanded] 
element.style { ()
    display: none;

[expanded] 
.foo { (styles-source-l…-inline.html:7 -> styles-source-lines-inline.html:7:7)
    color: blue;

[expanded] 
div { (user agent stylesheet)
/-- overloaded --/     display: block;
